Делит на модули и настраивает сборку
